Split dumpgenerator into files
