+++ wikiteam3/dumpgenerator/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-# dumpgenerator.py A generator of dumps for wikis
+# py A generator of dumps for wikis
 # Copyright (C) 2011-2018 WikiTeam developers
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -20,8 +20,8 @@
 #     https://github.com/WikiTeam/wikiteam/wiki
 
 
-
 from .generator import DumpGenerator
+
 
 def main():
     DumpGenerator()
