+++ wikiteam3/dumpgenerator/domain.py
@@ -1,0 +1,20 @@
+import re
+
+
+def domain2prefix(config={}, session=None):
+    """Convert domain name to a valid prefix filename."""
+
+    # At this point, both api and index are supposed to be defined
+    domain = ""
+    if config["api"]:
+        domain = config["api"]
+    elif config["index"]:
+        domain = config["index"]
+
+    domain = domain.lower()
+    domain = re.sub(r"(https?://|www\.|/index\.php.*|/api\.php.*)", "", domain)
+    domain = re.sub(r"/", "_", domain)
+    domain = re.sub(r"\.", "", domain)
+    domain = re.sub(r"[^A-Za-z0-9]", "_", domain)
+
+    return domain
