+++ wikiteam3/dumpgenerator/exceptions.py
@@ -1,0 +1,15 @@
+class PageMissingError(Exception):
+    def __init__(self, title, xml):
+        self.title = title
+        self.xml = xml
+
+    def __str__(self):
+        return "page '%s' not found" % self.title
+
+
+class ExportAbortedError(Exception):
+    def __init__(self, index):
+        self.index = index
+
+    def __str__(self):
+        return "Export from '%s' did not return anything." % self.index
