+++ wikiteam3/dumpgenerator/user_agent.py
@@ -1,0 +1,9 @@
+def getUserAgent():
+    """Return a cool user-agent to hide Python user-agent"""
+    useragents = [
+        # firefox
+        # 'Mozilla/5.0 (X11; Linux x86_64; rv:72.0) Gecko/20100101 Firefox/72.0',
+        # 'Mozilla/5.0 (X11; Linux x86_64; rv:68.0) Gecko/20100101 Firefox/68.0',
+        "Mozilla/5.0 (Windows NT 10.0; rv:78.0) Gecko/20100101 Firefox/78.0"
+    ]
+    return useragents[0]
