+++ wikiteam3/dumpgenerator/version.py
@@ -1,0 +1,5 @@
+__VERSION__ = "0.4.0-alpha"  # major, minor, micro: semver.org
+
+
+def getVersion():
+    return __VERSION__
