+++ wikiteam3/uploader.py
@@ -56,7 +56,7 @@ def log(wiki, dump, msg, config={}):
 
 
 def upload(wikis, config={}, uploadeddumps=[]):
-    headers = {"User-Agent": dumpgenerator.getUserAgent()}
+    headers = {"User-Agent": getUserAgent()}
     dumpdir = config.wikidump_dir
 
     filelist = os.listdir(dumpdir)
@@ -67,7 +67,7 @@ def upload(wikis, config={}, uploadeddumps=[]):
         wiki = wiki.lower()
         configtemp = config
         try:
-            prefix = dumpgenerator.domain2prefix(config={"api": wiki})
+            prefix = domain2prefix(config={"api": wiki})
         except KeyError:
             print("ERROR: could not produce the prefix for %s" % wiki)
         config = configtemp
@@ -343,7 +343,7 @@ The list must be a text file with the wiki's api.php URLs, one per line.
 Dumps must be in the same directory and follow the -wikidump.7z/-history.xml.7z format
 as produced by launcher.py (explained in https://github.com/WikiTeam/wikiteam/wiki/Tutorial#Publishing_the_dump ).
 You need a file named keys.txt with access and secret keys, in two different lines
-You also need dumpgenerator.py in the same directory as this script.
+You also need py in the same directory as this script.
 
 Use --help to print this help."""
     )
